[sk] Mode imputation works on all dtypes

+++ mage_ai/data_cleaner/transformer_actions/column.py
@@ -81,7 +81,7 @@ def impute(df, action, **kwargs):
     elif strategy == ImputationStrategy.MEDIAN:
         df[columns] = df[columns].fillna(df[columns].astype(float).median(axis=0))
     elif strategy == ImputationStrategy.MODE:
-        df[columns] = df[columns].fillna(df[columns].astype(float).mode(axis=0).iloc[0])
+        df[columns] = df[columns].fillna(df[columns].mode(axis=0).iloc[0])
     elif strategy == ImputationStrategy.COLUMN:
         replacement_df = pd.DataFrame({col: df[value] for col in columns})
         df[columns] = df[columns].fillna(replacement_df)
